feat: 증상 보고서 조회 및 검색 API 추가
- 각 병원에 제출된 증상 보고서를 조회하거나 검색 할 수 있는 API 입니다.
- 검색 필터링은 날짜, 증상, 중증도를 GET방식의 URL Query 형식으로 구분하며, 필터링을 걸지 않으면 전체 데이터가 조회됩니다.
- hospitals 및 reports 테이블의 위도, 경도 데이터를 소수점 4자리까지 밖에 저장할 수 없는 문제가 있어 해당 컬럼의 데이터 타입을 float에서 decimal로 변경하였습니다.

+++ src/hospitals/hospitals.entity.ts
@@ -25,10 +25,10 @@ export class Hospitals extends BaseEntity {
   @Column('int')
   available_beds: number;
 
-  @Column('float')
+  @Column('decimal')
   latitude: number;
 
-  @Column('float')
+  @Column('decimal')
   longitude: number;
 
   @CreateDateColumn()
